feat: 레시피 상세조회에 ingredient.calories 필드 추가

+++ src/main/java/com/jdc/recipe_service/domain/entity/Ingredient.java
@@ -33,7 +33,7 @@ public class Ingredient {
     private String unit;
 
     @Column(name = "calorie", nullable = true)
-    private Integer calorie;
+    private Double calorie;
 
     @Column(name = "english_name", length = 100)
     private String englishName;
